feat(bandit): dataset-steering reward → binary can_submit (submittable yield)
Task 3405's "0 PASS" root cause was the dataset-steering bandit reward steering
FLAT onto exhausted datasets. The reward is reworked to target SUBMITTABLE
yield, not edge — verified against live data: model16 has 110 sharpe≥1.25 alphas
but 0/110 can_submit (all fail CONCENTRATED_WEIGHT — sparse fscore fields
concentrate book weight), while pv1 has 61 can_submit. An edge reward points at
fool's gold.

Reward = binary `can_submit` (1.0/0.0). A graded edge term is dead under this
gate anyway (can_submit ⟹ sharpe≥1.25 ∧ fitness≥1.0 ⟹ any min ratio caps at
1.0), so the graded/convex apparatus retires. Fixes:
- B1: _aggregate adds `task_id IS NOT NULL` — the prior pure-column read would
  ingest ~4665 BRAIN-synced user alphas (they carry dataset_id + can_submit) and
  re-pollute the signal (v1's _iqc_marginal stamp had accidentally excluded them).
- first_run: re-seed via TRUNCATE (first_run = not states is GLOBAL; a scoped
  region DELETE would leave first_run False → seed from a 7-day window → collapse).
- H1: seed ALL active catalog datasets so none sits at the column default
  mining_weight=1.0 (which would dominate the bandit's sub-1.0 weights in
  ORDER BY / weighted_choice). Zero-history ones use a pessimistic cold-start
  Beta(1, COLDSTART_BETA) — the seed formula now branches on t_seed==0 (the
  1.0+max(0,t-s) form would always give Beta(1,1), not Beta(1,β)).
- selection_strategy.discounted_thompson_update: s_d int→float (fractional
  reward forward-compat); t_d stays int.
- Grid-tuned floor_c=0.03 / tau=100 / coldstart_beta=5 (can_submit rate ~0.1 is
  comparable to the floor, so a high floor would drown proven submitters).

Deployed (reversible): backup bandit_state + mining_weight + watermark, TRUNCATE,
re-seed. Verified model16 sinks to the bottom, analyst4/pv1 to the top, no
dataset at 1.0, pv1 pulls=574 (AIAC all-history, not synced-inflated). Reward
limitation: can_submit ≠ portfolio-marginal value. Next lever: FieldScreener is
not wired into FLAT (field/expression generation within a steered dataset).

+++ backend/selection_strategy.py
@@ -204,7 +204,7 @@ class DatasetBandit:
 # =============================================================================
 
 def discounted_thompson_update(
-    alpha: float, beta: float, s_d: int, t_d: int, gamma: float = 0.95
+    alpha: float, beta: float, s_d: float, t_d: int, gamma: float = 0.95
 ) -> Tuple[float, float]:
     """Pull-indexed discounted Beta-Bernoulli posterior update.
 
@@ -225,8 +225,8 @@ def discounted_thompson_update(
     equals (cumulative pulls − pulls_at_last_refresh): pull-indexed, not
     calendar — an arm that wasn't pulled this window has g=gamma**0=1.
     """
-    s_d = max(0, int(s_d))
-    t_d = max(0, int(t_d))
+    s_d = max(0.0, float(s_d))   # float: reward may be fractional (graded-reward forward-compat)
+    t_d = max(0, int(t_d))       # pull count stays integer (g=gamma**t_d exponent + pulls)
     s_d = min(s_d, t_d)  # clamp: s_d ⊆ t_d by construction; guard bad inputs
     g = gamma ** t_d
     new_alpha = g * float(alpha) + s_d
